+++ configs/supervisord.conf
@@ -5,13 +5,13 @@ nodaemon=true
 command=/bin/bash -c "cd /root/snort/snort3_monitor && python3 manage.py migrate && python3 manage.py runserver 0.0.0.0:8000"
 
 [program:watcher]
-command=/bin/bash -c "cd /root/snort/snort3_monitor && python3 script_rules.py; python3 script_watcher.py"
+command=/bin/bash -c "cd /root/snort/snort3_monitor && python3 update_rules.py; python3 watcher_alert.py"
 
 [program:performance]
-command=/bin/bash -c "cd /root/snort/snort3_monitor && python3 script_watcher_perf.py"
+command=/bin/bash -c "cd /root/snort/snort3_monitor && python3 watcher_perf.py"
 
 [program:snort]
-command=/bin/bash -c "snort -c /usr/local/etc/snort/snort.lua --plugin-path=/usr/local/etc/so_rules/ -s 65535 -k none -l /var/log/snort -i eth0 -m 0x1b --tweaks custom"
+command=/bin/bash -c "snort -c /usr/local/etc/snort/snort.lua --plugin-path=/usr/local/etc/so_rules/ -s 65535 -k none -l /var/log/snort -i eth0 -m 0x1b --tweaks custom --shell -j 12345"
 
 [program:cron]
 command=cron -f
