+++ snort3_monitor/clear_database.py
